+++ src/exactly_lib/cli/program_modes/common/argument_parsing_of_actor.py
@@ -1,7 +1,8 @@
-import shlex
 from typing import List, Optional
 
 from exactly_lib.actors.source_interpreter import actor
+from exactly_lib.cli.definitions import common_cli_options
+from exactly_lib.cli.program_modes.common.shlex_arg_parse import shlex_split
 from exactly_lib.definitions.entity import actors
 from exactly_lib.symbol.data import list_sdvs, string_sdvs
 from exactly_lib.symbol.logic.program.arguments_sdv import ArgumentsSdv
@@ -11,22 +12,16 @@ from exactly_lib.util.name_and_value import NameAndValue
 
 
 def resolve_actor_from_argparse_argument(default_actor: NameAndValue[Actor],
-                                         interpreter: List[str]) -> NameAndValue[Actor]:
-    interpreter_argument = None
-    if interpreter and len(interpreter) > 0:
-        interpreter_argument = interpreter[0]
-    return _resolve_act_phase_setup(default_actor, interpreter_argument)
+                                         interpreter: Optional[List[str]]) -> NameAndValue[Actor]:
+    if not interpreter:
+        return default_actor
+
+    interpreter_cmd_and_args = shlex_split(common_cli_options.OPTION_FOR_ACTOR__LONG, interpreter[0])
+
+    return _source_interpreter_with_system_command(interpreter_cmd_and_args)
 
 
-def _resolve_act_phase_setup(default_actor: NameAndValue[Actor],
-                             interpreter: Optional[str] = None) -> NameAndValue[Actor]:
-    if interpreter:
-        return _new_for_generic_script_language_setup(interpreter)
-    return default_actor
-
-
-def _new_for_generic_script_language_setup(interpreter: str) -> NameAndValue[Actor]:
-    cmd_and_args = shlex.split(interpreter)
+def _source_interpreter_with_system_command(cmd_and_args: List[str]) -> NameAndValue[Actor]:
     command_sdv = command_sdvs.for_system_program(
         string_sdvs.str_constant(cmd_and_args[0]),
         ArgumentsSdv(list_sdvs.from_str_constants(cmd_and_args[1:])),